Move server project, add client project
