+++ client/src/main/java/de/maik/reactivespringclient/ReactiveSpringClientApplication.java
@@ -1,0 +1,13 @@
+package de.maik.reactivespringclient;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReactiveSpringClientApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReactiveSpringClientApplication.class, args);
+	}
+
+}
