+++ client/src/test/java/de/maik/reactivespringclient/ReactiveSpringClientApplicationTests.java
@@ -1,0 +1,13 @@
+package de.maik.reactivespringclient;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ReactiveSpringClientApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
