+++ server/src/main/java/de/maik/reactivespring/server/ReactiveSpringApplication.java
@@ -1,0 +1,13 @@
+package de.maik.reactivespring.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReactiveSpringApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReactiveSpringApplication.class, args);
+	}
+
+}
