+++ server/src/main/java/de/maik/reactivespring/server/item/document/Item.java
@@ -1,0 +1,21 @@
+package de.maik.reactivespring.server.item.document;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+/**
+ * Inventory item representing stock
+ */
+@Document
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Item {
+    @Id
+    private String id;
+    private String description;
+    private double price;
+}
