+++ server/src/main/java/de/maik/reactivespring/server/item/document/ItemRepository.java
@@ -1,0 +1,9 @@
+package de.maik.reactivespring.server.item.document;
+
+import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
+import reactor.core.publisher.Flux;
+
+public interface ItemRepository extends ReactiveMongoRepository<Item, String> {
+
+    Flux<Item> findByDescription(String description);
+}
