--- src/main/java/de/maik/reactivespring/ReactiveSpringApplication.java
@@ -1,13 +1,0 @@
-package de.maik.reactivespring;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReactiveSpringApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ReactiveSpringApplication.class, args);
-	}
-
-}
