--- src/main/java/de/maik/reactivespring/item/document/ItemRepository.java
@@ -1,9 +1,0 @@
-package de.maik.reactivespring.item.document;
-
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-import reactor.core.publisher.Flux;
-
-public interface ItemRepository extends ReactiveMongoRepository<Item, String> {
-
-    Flux<Item> findByDescription(String description);
-}
